requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
aiohttp==3.10.8
python-dotenv==1.0.1
pydantic~=2.9.2
httpx==0.27.2